refactor: extract getAuthHeaders into shared auth.ts utility
